chore: make docs.rs build all features
Specify to docs.rs that it should build with all features. This will
result in docs.rs compiling and documenting the `range` module, as well
as showing that the `Isbn*` types support serde's `Deserialize` and
`Serialize` traits. The `doc_cfg` feature makes rustdoc render what
features are required for something. Images in merge request.